Move time helpers back into time scale
